tests/wpt/tests/html/semantics/forms/the-form-element/form-elements-interfaces-01.html

<!DOCTYPE html>
<title>form.elements: interfaces</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-form-elements">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#htmlformcontrolscollection">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var form = document.createElement("form");
  ["HTMLFormControlsCollection", "HTMLCollection"].forEach(function(i) {
    test(function() {
      assert_true(i in window, "Interface should exist")
      assert_true(form.elements instanceof window[i],
                  "elements should implement the interface")
    }, "Testing interface " + i)
  })
})
</script>
